Temp revert back to non-kotlin swift java coder; expand java bridgeable collection extensions to support dictionary
