+++ Package.swift
@@ -13,7 +13,7 @@ let package = Package(
 	dependencies: [
 		.package(url: "https://github.com/readdle/java_swift", .upToNextMinor(from: "2.1.9")),
 		.package(url: "https://github.com/readdle/swift-java", .upToNextMinor(from: "0.2.4")),
-		.package(url: "https://github.com/readdle/swift-java-coder", .branch("dev/kotlin-support")),
+		.package(url: "https://github.com/readdle/swift-java-coder", .upToNextMinor(from: "1.0.17")),
 		.package(url: "https://github.com/readdle/swift-anycodable.git", .upToNextMinor(from: "1.0.3"))
 	],
 	targets: [
